tests/integration/requirements.txt
# Integration test dependencies
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
requests>=2.31.0
PyYAML>=6.0
